% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statestack.R
\name{is.statestack}
\alias{is.statestack}
\title{Is StateStack}
\usage{
is.statestack(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class \code{\link{StateStack}}.
}

