% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statestack.R
\docType{data}
\name{StateStack}
\alias{StateStack}
\title{Stack Classs for storing a stack of states class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StateStack
}
\description{
Stack Classs for storing a stack of states class
}
\section{Fields}{

\describe{
\item{\code{state}}{allows the user to peek or set the top value of the stack}
}}
\keyword{datasets}

